Refactor with get_repo_path function inside enum
